pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mltrack"
version = "0.1.0"
description = "Universal ML tracking tool for teams"
readme = "README.md"
authors = [
    { name = "Your Team", email = "ml-team@yourcompany.com" }
]
license = "MIT"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8"
dependencies = [
    "mlflow>=2.0.0",
    "click>=8.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "gitpython>=3.1",
    "pyyaml>=6.0",
    "psutil>=5.9",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "torchaudio>=2.5.1",
    "tqdm>=4.67.1",
    "openai>=1.93.2",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.3.2",
    "anthropic>=0.57.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
]
sklearn = ["scikit-learn>=1.0"]
torch = ["torch>=2.0"]
tensorflow = ["tensorflow>=2.0"]
all = ["mltrack[sklearn,torch,tensorflow]"]

[project.scripts]
mltrack = "mltrack.cli:main"

[project.urls]
"Homepage" = "https://github.com/yourorg/mltrack"
"Bug Reports" = "https://github.com/yourorg/mltrack/issues"
"Source" = "https://github.com/yourorg/mltrack"

[tool.hatch.build.targets.wheel]
packages = ["src/mltrack"]

[tool.ruff]
target-version = "py38"
line-length = 100
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "SIM", "ARG"]

[tool.mypy]
python_version = "3.8"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=mltrack --cov-report=term-missing"

[tool.coverage.run]
source = ["src/mltrack"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
    "ipykernel>=6.0", # For notebook demos
    "toml>=0.10.2",
]
